+++ src/packages/menu/CMenu.tsx
@@ -22,6 +22,7 @@ export default defineComponent({
       read: {},
       state: {
         menuList,
+        active,
       },
       style: {
         classList: {